update view folder, style Structural unification
