+++ samples/src/components/View/Content.js
@@ -1,7 +1,6 @@
 /* eslint-disable */
 import React from 'react';
 import Container from '@material-ui/core/Container';
-import { makeStyles } from '@material-ui/core/styles';
 import IBSheet8 from 'components/Create/SheetCreate';
 import DialogSheet from 'samples/Dialog/component';
 import MergeFunction from 'samples/Merge/function';
@@ -11,6 +10,7 @@ import MultiFunction from 'samples/Multi/function';
 import TreeFunction from 'samples/Tree/function';
 import FormDiv from 'samples/Form/component';
 import { useSelector } from 'react-redux';
+import { contentClasses } from './Features/ViewStyle';
 
 // 각 샘플 컴포넌트에서 title, subTitle, func 받아오는 것은 sheet 컴포넌트 쪽 탭을 만들어서 사용함.
 const Content = () => {
@@ -18,39 +18,7 @@ const Content = () => {
   const subTitle = useSelector((state) => state.subTitle);
   const name = useSelector((state) => state.name);
   const sheet = useSelector((state) => state.sheet);
-
-  const useStyles = makeStyles((props) => ({
-    content: {
-      fontFamily: 'Noto Sans CJK KR,sans-serif;',
-      margin: '0px auto',
-      maxWidth: '1200px',
-      paddingLeft: '0px',
-      paddingRight: '0px'
-    },
-    mainDiv: {
-      paddingTop: '15px',
-      width: '100%',
-      paddingLeft: '24px',
-      paddingLight: '24px'
-    },
-    subDiv: {
-      width: '100%'
-    },
-    title: {
-      fontSize: '1.1rem',
-      fontWeight: 'bold',
-      color: '#000'
-    },
-    subTitle: {
-      color: '#4c4c57',
-      fontSize: '1rem'
-    },
-    divRow: {
-      display: 'flex'
-    }
-  }));
-
-  const classes = useStyles();
+  const classes = contentClasses();
 
   return (
     <>
