--- samples/src/components/View/top.module.css
@@ -1,19 +1,0 @@
-.topBtn {
-  position: fixed;
-  opacity: 0;
-  bottom: 40px;
-  right: 40px;
-  z-index: 255;
-  width: 50px;
-  height: 50px;
-  border-radius: 100%;
-  border: none;
-  background: cornflowerblue;
-  color: white;
-  font-size: 18px;
-  font-weight: bold;
-  letter-spacing: -0.06em;
-  box-shadow: 1px 1px 6px 3px rgba(0,0,0,0.3);
-  cursor: pointer;
-  transition: opacity 0.3s ease-in;
-}
